Rename db models
